requirements.txt
ebooklib
pillow
